feat(api): add POST route for creating a store
A new file `route.ts` was added to the `app/api/stores` directory. This file contains the implementation of the `POST` route for creating a store. The route expects a JSON payload with a `name` field. If the user is not authenticated, a `401 Unauthorized` response is returned. If the `name` field is missing, a `400 Bad Request` response is returned. Otherwise, the store is created in the database and a JSON response with the created store is returned. In case of any errors, a `500 Internal Server Error` response is returned.

fix(layout): import and use `prismadb` in `layout.tsx`

In the `app/layout.tsx` file, the `prismadb` module was imported from the `@/lib/prismadb` directory. This module is used to access the `store` object. This change allows the `store` object to be used in the layout component.

feat(modals): add loading state and toast notifications to `store-modal.tsx`

In the `components/modals/store-modal.tsx` file, the `useForm` hook was imported from the `react-hook-form` library. The `axios` library was imported for making HTTP requests. The `set` function from the `react-hook-form` library was imported to update form values. The `useState` hook was imported from the `react` library to manage the loading state. The `toast` function from the `react-hot-toast` library was imported for displaying toast notifications.

The `loading` state was added to manage the loading state of the form submission. When the form is submitted, the loading state is set to `true`. An HTTP POST request is made to the `/api/stores` endpoint with the form values. If the request is successful, a success toast notification is displayed. If there is an error, an error toast notification is displayed. Finally, the loading state is set to `false`.

feat(prismadb): add `prismadb` module

A new file `prismadb.ts` was added to the `lib` directory. This file exports a `prismadb` object which is an instance of the `PrismaClient` class from the `@prisma/client` library. The `prismadb` object is assigned to the global `prisma` variable. If the `NODE_ENV`